Move LineMarkerPosition interface to types
