+++ src/components/editor/document-renderer-pane/document-render-pane.tsx
@@ -4,7 +4,7 @@ import useResizeObserver from 'use-resize-observer'
 import { TocAst } from '../../../external-types/markdown-it-toc-done-right/interface'
 import { ForkAwesomeIcon } from '../../common/fork-awesome/fork-awesome-icon'
 import { ShowIf } from '../../common/show-if/show-if'
-import { LineMarkerPosition } from '../../markdown-renderer/basic-markdown-renderer'
+import { LineMarkerPosition } from '../../markdown-renderer/utils/types'
 import { FullMarkdownRenderer } from '../../markdown-renderer/full-markdown-renderer'
 import { ScrollProps, ScrollState } from '../scroll/scroll-props'
 import { findLineMarks } from '../scroll/utils'
