+++ src/components/markdown-renderer/full-markdown-renderer.tsx
@@ -17,7 +17,7 @@ import markdownItRegex from 'markdown-it-regex'
 import subscript from 'markdown-it-sub'
 import superscript from 'markdown-it-sup'
 import toc from 'markdown-it-toc-done-right'
-import React, { Fragment, useCallback, useMemo, useRef, useState } from 'react'
+import React, { useCallback, useMemo, useRef, useState } from 'react'
 import { Alert } from 'react-bootstrap'
 import { Trans } from 'react-i18next'
 import { useSelector } from 'react-redux'
@@ -27,7 +27,7 @@ import { InternalLink } from '../common/links/internal-link'
 import { ShowIf } from '../common/show-if/show-if'
 import { slugify } from '../editor/table-of-contents/table-of-contents'
 import { RawYAMLMetadata, YAMLMetaData } from '../editor/yaml-metadata/yaml-metadata'
-import { AdditionalMarkdownRendererProps, BasicMarkdownRenderer, LineMarkerPosition } from './basic-markdown-renderer'
+import { AdditionalMarkdownRendererProps, BasicMarkdownRenderer } from './basic-markdown-renderer'
 import { createRenderContainer, validAlertLevels } from './markdown-it-plugins/alert-container'
 import { combinedEmojiData } from './markdown-it-plugins/emoji/mapping'
 import { firstHeaderExtractor } from './markdown-it-plugins/first-header-extractor'
@@ -67,6 +67,7 @@ import { TaskListReplacer } from './replace-components/task-list/task-list-repla
 import { VimeoReplacer } from './replace-components/vimeo/vimeo-replacer'
 import { YoutubeReplacer } from './replace-components/youtube/youtube-replacer'
 import { useCalculateLineMarkerPosition } from './utils/calculate-line-marker-positions'
+import { LineMarkerPosition } from './utils/types'
 import { usePostMetaDataOnChange } from './utils/use-post-meta-data-on-change'
 import { usePostTocAstOnChange } from './utils/use-post-toc-ast-on-change'
