+++ src/components/markdown-renderer/utils/calculate-line-marker-positions.ts
@@ -1,7 +1,7 @@
 import equal from 'fast-deep-equal'
 import { RefObject, useCallback, useEffect, useRef } from 'react'
 import useResizeObserver from 'use-resize-observer'
-import { LineMarkerPosition } from '../basic-markdown-renderer'
+import { LineMarkerPosition } from './types'
 import { LineMarkers } from '../markdown-it-plugins/line-number-marker'
 
 export const calculateLineMarkerPositions = (documentElement: HTMLDivElement, currentLineMarkers: LineMarkers[], offset?: number): LineMarkerPosition[] => {
