+++ src/components/markdown-renderer/utils/types.d.ts
@@ -2,3 +2,8 @@ export interface LineKeys {
   line: string,
   id: number
 }
+
+export interface LineMarkerPosition {
+  line: number
+  position: number
+}
